old-tests/submission/W3C/domhtml/HTMLAnchorElement01.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Anchor</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/editing.html#dom-accesskey" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<a accesskey="g" charset="US-ASCII" coords="0,0,100,100" dir="LTR" href="/support/images/yellow-rectangle.png" hreflang="en" id="Anchor" name="Anchor" rel="GLOSSARY" rev="STYLESHEET" shape="rect" tabindex="22" type="image/png">View Submit Button</a>
</p>


<script>
var domtest = async_test("The accessKey attribute is a single character access key to give     access to the form control.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vaccesskey;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("a");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vaccesskey = testNode.accessKey;

      assert_equals(vaccesskey,"g");
       
    domtest.done();
}));

</script>
<p>The accessKey attribute is a single character access key to give
    access to the form control. 

    Retrieve the accessKey attribute and examine its value.</p>
<div id="log"></div>
</body></html>
